Using hwloc to detect topology. And some cleanup.

+++ capacity.hpp
@@ -3,30 +3,30 @@
 #include <iostream>
 
 struct capacity {
-	bool enough(int v) { return _value >= v; }
+	bool enough(size_t v) { return _value >= v; }
 
-	capacity &operator=(const int c) {
+	capacity &operator=(const size_t c) {
 		_value = c;
 		return *this;
 	}
 
-	capacity &operator-=(const int c) {
+	capacity &operator-=(const size_t c) {
 		_value -= c;
 		return *this;
 	}
 
-	capacity &operator+=(const int c) {
+	capacity &operator+=(const size_t c) {
 		_value += c;
 		return *this;
 	}
 
-	capacity &operator*=(const int c) {
+	capacity &operator*=(const size_t c) {
 		_value *= c;
 		return *this;
 	}
 
   private:
-	int _value;
+	size_t _value;
 
 	friend std::ostream &operator<<(std::ostream &o, const capacity &p);
 };
